Adds Player and ir module.
