+++ src/lcgs/ir/mod.rs
@@ -1,0 +1,3 @@
+mod intermediate;
+mod eval;
+mod symbol_table;
