+++ src/lcgs/mod.rs
@@ -1,5 +1,4 @@
 mod ast;
-mod symbol_table;
 mod parse;
 mod precedence;
-mod intermediate;
+mod ir;
